requirements.txt
tabulate==0.8.7
psycopg2==2.9.6
python-dotenv==1.0.0
unicodecsv==0.14.1